refactor: home page
* Updated routing to allow home page to have different layout.tsx to all the other pages
* Reponsiveness for splash header
* Responsiveness for navigation (hamburger menu)
