+++ package.json
@@ -15,9 +15,9 @@
     "emailjs-com": "^3.2.0",
     "framer-motion": "^10.0.0",
     "mongoose": "^8.2.3",
-    "next": "14.1.1",
-    "react": "^18",
-    "react-dom": "^18",
+    "next": "^14.2.4",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "three": "^0.164.1",
     "three-orbitcontrols": "^2.110.3"
   },
@@ -30,7 +30,7 @@
     "autoprefixer": "^10.0.1",
     "daisyui": "^4.7.3",
     "eslint": "8.57.0",
-    "eslint-config-next": "14.1.4",
+    "eslint-config-next": "^14.2.4",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
     "postcss": "^8",
