+++ src/app/(home)/page.tsx
@@ -1,18 +1,14 @@
 "use client";
 
 import Navigation from "@/components/Navigation";
-import Footer from "@/components/Footer";
 import SplashHeader from "@/components/SplashHeader";
 
-import HomeBackground from "../../public/homeBackgroundGreen.png";
-import TestImage from "../../public/test_image.jpeg";
+import HomeBackground from "../../../public/homeBackgroundGreen.png";
 
 // below are possible ways to add the coloured overlay using css (which would make it easy to change themes)
 // currently the green background is part odf the image because text appears behind the coloured overlay otherwise.
 // , boxShadow: "inset 0 0 0 1000px rgba(0,55,0,.2)"
 // <div className="fixed top-0 left-0 w-48 h-full z-0 bg-base-100 z-0 opacity-80"></div>
-
-const images: string[] = [HomeBackground.src, TestImage.src];
 
 export default function Home() {
   return (
@@ -20,7 +16,6 @@ export default function Home() {
       <div
         className="bg-cover bg-center"
         style={{ backgroundImage: `url(${HomeBackground.src})` }}
-        // style={{ backgroundImage: `url(${HomeBackground.src})` }}
       >
         <div className="h-screen flex flex-col min-h-inherit">
           <div className="mx-auto max-w-screen-xl">
@@ -36,7 +31,6 @@ export default function Home() {
       <div>
         <p>About us</p>
       </div>
-      <Footer />
     </main>
   );
 }
