+++ src/app/(pages)/contact/page.tsx
@@ -1,0 +1,9 @@
+import ContactForm from "@/components/ContactForm";
+
+export default function JoinPage() {
+  return (
+    <div>
+      <ContactForm />
+    </div>
+  );
+}
