+++ src/app/(pages)/event-calendar/page.tsx
@@ -1,0 +1,9 @@
+import EventCalendar from "@/components/EventCalendar";
+
+export default function JoinPage() {
+  return (
+    <div>
+      <EventCalendar />
+    </div>
+  );
+}
