+++ src/app/(pages)/join/page.tsx
@@ -1,0 +1,3 @@
+export default function JoinPage() {
+  return <div></div>;
+}
