+++ src/app/(pages)/layout.tsx
@@ -2,7 +2,9 @@
 
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import "./globals.css";
+import "../globals.css";
+import Navigation from "@/components/Navigation";
+import Footer from "@/components/Footer";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -19,7 +21,13 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en" data-theme="sesaTheme">
-      <body className={inter.className}>{children}</body>
+      <body className={inter.className}>
+        <div className="bg-primary">
+          <Navigation />
+        </div>
+        {children}
+        <Footer />
+      </body>
     </html>
   );
 }
