+++ src/app/(pages)/our-team/page.tsx
@@ -1,0 +1,3 @@
+export default function OurTeamPage() {
+  return <div></div>;
+}
