+++ src/app/(pages)/sponsors/page.tsx
@@ -1,0 +1,3 @@
+export default function SponsorsPage() {
+  return <div></div>;
+}
