--- src/app/join/page.tsx
@@ -1,14 +1,0 @@
-import Navigation from "@/components/Navigation";
-import Footer from "@/components/Footer";
-
-export default function JoinPage() {
-  return (
-    <div>
-      <div className="bg-primary">
-        <Navigation />
-      </div>
-      
-      <Footer />
-    </div>
-  );
-}
