--- src/app/our-team/page.tsx
@@ -1,14 +1,0 @@
-import Navigation from "@/components/Navigation";
-import Footer from "@/components/Footer";
-
-export default function OurTeamPage() {
-  return (
-    <div>
-      <div className="bg-primary">
-        <Navigation />
-      </div>
-
-      <Footer />
-    </div>
-  );
-}
