+++ src/components/SplashHeader.tsx
@@ -3,7 +3,6 @@ import ThreeScene from "./ThreeScene";
 import SocialMediaButtons from "./SocialMediaButtons";
 
 interface SplashHeaderProps {
-  images: string[];
   headerText: string;
   bodyText: string;
 }
@@ -14,9 +13,9 @@ const SplashHeader: React.FC<SplashHeaderProps> = ({
 }) => {
   return (
     <div className="relative mx-auto py-32">
-      <div className="flex items-center">
+      <div className="flex items-center px-5">
         <div className="w-2/3">
-          <h1 className="text-accent text-4xl sm:text-5xl md:text-6xl xl:text-7xl text-shadow font-bold">
+          <h1 className="text-accent text-5xl md:text-6xl xl:text-7xl text-shadow font-bold">
             {headerText}
           </h1>
           <p className="text-white md:text-xl xl:text-2xl text-shadow pt-2">{bodyText}</p>
@@ -25,7 +24,7 @@ const SplashHeader: React.FC<SplashHeaderProps> = ({
           </div>
         </div>
       </div>
-      <div className="absolute top-0 right-0 h-full w-1/2 p-0 flex items-center overflow-hidden">
+      <div className="hidden lg:flex absolute top-0 right-0 h-full w-1/2 p-0 items-center overflow-hidden">
         <ThreeScene />
       </div>
     </div>
